fix(ci): get lint, typecheck, and test jobs green
- Switch CI lint from black --check to ruff format --check and align
  the repo on ruff format (kills the recurring ruff/black drift that
  kept re-reddening CI; matches the local pre-commit hook).
- Resolve all ruff check violations: unused imports/vars, overlong
  lines, import sorting, ambiguous `l` variable.
- Resolve mypy strict errors: float cast in NewsBelief, dict/list type
  params, __all__ re-exports for the legacy shims, Callable/Any
  annotations, multiprocessing BaseProcess typing; targeted
  # type: ignore for object-typed Nautilus/VADER interop handles.

+++ src/qts/strategies/belief.py
@@ -44,7 +44,7 @@ class NewsBelief:
         half_life_s = self.half_life.total_seconds()
         if half_life_s <= 0:
             return 1.0
-        return 0.5 ** (elapsed / half_life_s)
+        return float(0.5 ** (elapsed / half_life_s))
 
     def observe(self, signal: NewsSignal, now: datetime) -> None:
         """Fold a classified signal into the belief at time `now`."""
